% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDist.R
\name{DataDist}
\alias{DataDist}
\title{modify datadist}
\usage{
DataDist(
  ...,
  data,
  q.display,
  q.effect = c(0.25, 0.75),
  adjto.cat = c("mode", "first"),
  n.unique = 10
)
}
\arguments{
\item{...}{see datadist\link{rms}}

\item{data}{see datadist\link{rms}}

\item{q.display}{see datadist\link{rms}}

\item{q.effect}{see datadist\link{rms}}

\item{adjto.cat}{see datadist\link{rms}}

\item{n.unique}{see datadist\link{rms}}
}
\value{
see datadist\link{rms}
}
\description{
modify datadist
}
