% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChangepoints.R
\name{getChangepoints}
\alias{getChangepoints}
\title{getChangepoints}
\usage{
getChangepoints(r, range = NULL)
}
\arguments{
\item{r}{result of RCS()}

\item{range}{range for the rcs() varaibel}
}
\value{
dataframe
}
\description{
getChangepoints
}
