% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Referenceget.R
\name{getReference}
\alias{getReference}
\title{adjsut reference for rcs() variable}
\usage{
getReference(fit, data = NULL)
}
\arguments{
\item{fit}{result of cph(), lrm() or ols()}
}
\value{
dataframe
}
\description{
adjsut reference for rcs() variable
}
\details{
\itemize{
\item for cph() and lrm(), value of x whith yhat close to 1
\item for ols(), value of x whith yhat close to 0
}
}
