% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.m1by.R, R/ggplot_m1.R
\name{ggplot.m1by}
\alias{ggplot.m1by}
\alias{ggplot.m1}
\title{Plot for decision curve}
\usage{
\method{ggplot}{m1by}(
  data,
  mapping,
  ...,
  environment = parent.frame(),
  vline = TRUE,
  hline = TRUE,
  point = FALSE
)

\method{ggplot}{m1}(
  data,
  mapping,
  ...,
  environment = parent.frame(),
  vline = TRUE,
  hline = TRUE,
  point = FALSE
)
}
\arguments{
\item{data}{resultes of dca() function}

\item{mapping}{ignore}

\item{...}{ignore}

\item{environment}{ignore}

\item{color}{logical or colors}

\item{linetype}{logical or integers}

\item{lwd}{logical or integers}
}
\description{
Plot for decision curve

Plot for decision curve
}
