% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrcs2.R
\name{ggrcs2}
\alias{ggrcs2}
\title{ggrcs2}
\arguments{
\item{data}{need a dataframe}

\item{fit}{You need the fitted model.Must be  lrm or coxph.}

\item{x}{The target variable you wish to fit. It is displayed on the X-axis when plotting.}
}
\value{
a picture
}
\description{
A Function to Draw Histograms and Restricted Cubic Splines (RCS)
}
\details{
You can use this function to easily draw a combined.histogram and restricted cubic spline.The function draws the graph through ggplot2.RCS fitting requires the use of the rcs function of the RMS package.Can fit cox regression,logistic regression and linear regression models.
}
