% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knotoptimal.R
\name{optimalKnot}
\alias{optimalKnot}
\title{optimal Knot for rcs() function}
\usage{
optimalKnot(fit, n = 3:10, plot = TRUE, title = NULL, data = NULL)
}
\arguments{
\item{fit}{result of fit from 'rms' package}

\item{n}{number of nots}

\item{plot}{logical. whether to plot}
}
\value{
AIC and plot
}
\description{
optimal Knot for rcs() function
}
