% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predata.coxph.R
\name{predata.coxph}
\alias{predata.coxph}
\title{predata.coxph}
\usage{
\method{predata}{coxph}(fit, variables, y, group = NULL)
}
\arguments{
\item{fit}{Model function required for prediction.}

\item{variables}{variable name.}

\item{y}{the value of the variable.}

\item{group}{Variables that need to be grouped.}
}
\value{
Data required for plotting.
}
\description{
Generate the predicted data for the function. This is needed for drawing.
}
