% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Referenceset.R
\name{setReference}
\alias{setReference}
\title{set adjustment value}
\usage{
setReference(x, value)
}
\arguments{
\item{x}{variable}

\item{value}{value}
}
\value{
data distribution
}
\description{
set adjustment value
}
