% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlercs.R
\name{singlercs}
\alias{singlercs}
\title{singlercs}
\arguments{
\item{data}{need a dataframe}

\item{fit}{You need the fitted model.Must be  lrm , ols or coxph.}

\item{x}{The target variable you wish to fit. It is displayed on the X-axis when plotting.}
}
\value{
a picture
}
\description{
A Function to Draw Restricted Cubic Splines (RCS)
}
\details{
You can use this function to easily draw a restricted cubic spline.The function draws the graph through ggplot2.RCS fitting requires the use of the rcs function of the RMS package.Can fit cox regression,logistic regression and linear regression models.
}
\examples{
library(rms)
library(ggplot2)
library(scales)
dt<-smoke
dd<-datadist(dt)
options(datadist='dd')
fit<- cph(Surv(time,status==1) ~ rcs(age,4)+gender, x=TRUE, y=TRUE,data=dt)
###one group
singlercs(data=dt,fit=fit,x="age")
##two groups
singlercs(data=dt,fit=fit,x="age",group="gender")
}
