% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.R
\docType{data}
\name{smoke}
\alias{smoke}
\title{A data on age and smoking rates.}
\format{
An object of class \code{data.frame} with 995 rows and 5 columns.
}
\usage{
data(smoke)
}
\description{
A data on age and smoking rates.
}
\examples{
data(smoke)
}
\keyword{datasets}
