% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knotupdate.R
\name{updateKnot}
\alias{updateKnot}
\title{update Knot}
\usage{
updateKnot(fit, k, data = NULL)
}
\arguments{
\item{fit}{fit}

\item{k}{one number}
}
\value{
updated regression with new knot
}
\description{
update Knot
}
